refactoring packages more, before PR
