+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
 
         <service android:name=".main.util.SampleForegroundService" />
 
-        <activity android:name=".main.view.MainActivity"
+        <activity android:name=".main.MainActivity"
             android:configChanges="uiMode">
 
             <intent-filter>
