--- app/src/main/java/com/example/sampleappbyme/main/mvvm/AddEditTaskViewModel.kt
@@ -1,137 +1,0 @@
-/*
- * Copyright (C) 2019 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.sampleappbyme.main.mvvm
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import com.example.sampleappbyme.R
-import com.example.sampleappbyme.main.data.Event
-import com.example.sampleappbyme.main.data.SampleData
-import com.example.sampleappbyme.main.data.SampleDataSource
-import com.example.sampleappbyme.main.data.SampleRepository
-
-/**
- * ViewModel for the Add/Edit screen.
- *
- *
- * This ViewModel only exposes [ObservableField]s, so it doesn't need to extend
- * [androidx.databinding.BaseObservable] and updates are notified automatically. See
- * [com.example.android.architecture.blueprints.todoapp.statistics.StatisticsViewModel] for
- * how to deal with more complex scenarios.
- */
-class AddEditTaskViewModel(
-    private val tasksRepository: SampleRepository
-) : ViewModel() {
-
-    // Two-way databinding, exposing MutableLiveData
-    val title = MutableLiveData<String>()
-
-    // Two-way databinding, exposing MutableLiveData
-    val description = MutableLiveData<String>()
-
-    private val _dataLoading = MutableLiveData<Boolean>()
-    val dataLoading: LiveData<Boolean>
-        get() =_dataLoading
-
-
-    private val _snackbarText = MutableLiveData<Event<Int>>()
-    val snackbarMessage: LiveData<Event<Int>>
-        get() = _snackbarText
-
-    private val _taskUpdated = MutableLiveData<Event<Unit>>()
-    val taskUpdatedEvent: LiveData<Event<Unit>>
-        get() = _taskUpdated
-
-    private var taskId: String? = null
-
-    private var isNewTask: Boolean = false
-
-    private var isDataLoaded = false
-
-    private var taskCompleted = false
-
-    fun start(taskId: String?) {
-        _dataLoading.value?.let { isLoading ->
-            // Already loading, ignore.
-            if (isLoading) return
-        }
-        this.taskId = taskId
-        if (taskId == null) {
-            // No need to populate, it's a new task
-            isNewTask = true
-            return
-        }
-        if (isDataLoaded) {
-            // No need to populate, already have data.
-            return
-        }
-        isNewTask = false
-        _dataLoading.value = true
-
-//        tasksRepository.getTask(taskId, this)
-    }
-
-//    override fun onTaskLoaded(task: Task) {
-//        title.value = task.title
-//        description.value = task.description
-//        taskCompleted = task.isCompleted
-//        _dataLoading.value = false
-//        isDataLoaded = true
-//    }
-//
-//    override fun onDataNotAvailable() {
-//        _dataLoading.value = false
-//    }
-
-    // Called when clicking on fab.
-    internal fun saveTask() {
-//        val currentTitle = title.value
-//        val currentDescription = description.value
-//
-//        if (currentTitle == null || currentDescription == null) {
-//            _snackbarText.value =  Event(R.string.empty_task_message)
-//            return
-//        }
-//        if (SampleData(currentTitle, currentDescription ?: "").isEmpty) {
-//            _snackbarText.value =  Event(R.string.empty_task_message)
-//            return
-//        }
-//
-//        val currentTaskId = taskId
-//        if (isNewTask || currentTaskId == null) {
-//            createTask(Task(currentTitle, currentDescription))
-//        } else {
-//            val task = Task(currentTitle, currentDescription, currentTaskId)
-//                .apply { isCompleted = taskCompleted }
-//            updateTask(task)
-//        }
-    }
-
-//    private fun createTask(newTask: Task) {
-//                tasksRepository.saveTask(newTask)
-//                _taskUpdated.value = Event(Unit)
-//    }
-//
-//    private fun updateTask(task: Task) {
-//        if (isNewTask) {
-//            throw RuntimeException("updateTask() was called but task is new.")
-//        }
-//        tasksRepository.saveTask(task)
-//        _taskUpdated.value = Event(Unit)
-//    }
-}
