+++ app/src/main/java/com/example/sampleappbyme/main/view/fragment/TaskViewFragment.kt
@@ -1,26 +1,17 @@
 package com.example.sampleappbyme.main.view.fragment
 
-import android.opengl.Visibility
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.cardview.widget.CardView
 import androidx.lifecycle.Observer
-import androidx.viewpager.widget.ViewPager
 import com.example.sampleappbyme.R
 
-import com.example.sampleappbyme.databinding.CardViewFragmentBinding
 import com.example.sampleappbyme.databinding.TaskViewFragmentBinding
 import com.example.sampleappbyme.main.data.Event
-import com.example.sampleappbyme.main.ui.card.CardAdapter
-import com.example.sampleappbyme.main.ui.card.CardItem
-import com.example.sampleappbyme.main.ui.card.CardPagerAdapter
-import com.example.sampleappbyme.main.view.MainActivity
+import com.example.sampleappbyme.main.MainActivity
 import com.google.android.material.floatingactionbutton.FloatingActionButton
-import com.google.android.material.snackbar.Snackbar
 import timber.log.Timber
 
 class TaskViewFragment : Fragment() {
