+++ app/src/main/java/com/example/sampleappbyme/main/viewmodel/ViewModelFactory.kt
@@ -1,4 +1,4 @@
-package com.example.sampleappbyme.main
+package com.example.sampleappbyme.main.viewmodel
 
 import android.annotation.SuppressLint
 import android.app.Application
@@ -6,8 +6,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.example.sampleappbyme.main.data.SampleRepository
 import com.example.sampleappbyme.main.util.Injection
-import com.example.sampleappbyme.main.viewmodel.MainViewModel
-import com.example.sampleappbyme.main.viewmodel.TaskViewModel
 
 class ViewModelFactory private constructor(
     private val sampleRepository: SampleRepository
@@ -31,9 +29,12 @@ class ViewModelFactory private constructor(
         @Volatile private var INSTANCE: ViewModelFactory? = null
 
         fun getInstance(application: Application) =
-            INSTANCE ?: synchronized(ViewModelFactory::class.java) {
-                INSTANCE ?: ViewModelFactory(
-                    Injection.provideTasksRepository(application.applicationContext))
+            INSTANCE
+                ?: synchronized(ViewModelFactory::class.java) {
+                INSTANCE
+                    ?: ViewModelFactory(
+                        Injection.provideTasksRepository(application.applicationContext)
+                    )
                         .also { INSTANCE = it }
             }
     }
