+++ app/src/main/res/layout/addtask_act.xml
@@ -22,7 +22,7 @@
 
         <variable
             name="viewmodel"
-            type="com.example.sampleappbyme.main.mvvm.AddEditTaskViewModel" />
+            type="com.example.sampleappbyme.main.viewmodel.AddEditTaskViewModel" />
     </data>
 
     <androidx.drawerlayout.widget.DrawerLayout
